Remove task to copy shipped shipments to cartons
This was supposed to be ran back in the 1.2 days, and we don't need to
keep it around anymore. It'll still be in the repo history if anyone
needs it for whatever reason.

This also removes the migration, as anyone that has ran migrations on
any store in the last 6 versions will have it, and it stops migrations
in test_app's from running.