Add Dispatch
The Dispatcher is to allow for sending named messages to any named
messages to any number of recipients. Prior code would directly call
classes to email orders or promotions. Now, a class can just call the
dispatcher with a message name, and it will retrieve all of what it
needs to do from configuration, then sequentially call each configured
action.